refactor: remove screen components from common components
