+++ src/screens/main/styles.ts
